# src/main/resources/apiDocumentation.yaml
swagger: '2.0'
info:
  title: "KeukenTafelMeta API"
  description: "A web application for looking up Magic: The Gathering cards, make your own decklists and see decklists of friends."
  version: '2.0'
host: "keukentafelmeta.nl"
basePath: "/api"
tags:
  - name: "user"
    description: "Operations about user"
schemes:
  - "https"
paths:
  /user:
    post:
      tags:
        - "user"
      summary: "Create user"
      description: "Create a new account for an user with name, email, username and password."
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Create user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid email address supplied" #Is this the correct responses when an email address is already taken?
  /user/login:
    get:
      tags:
        - "user"
      summary: "Logs user into the system"
      description: ""
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: "username"
          in: "query"
          description: "The username for login"
          required: true
          type: "string"
        - name: "password"
          in: "query"
          description: "The password for login in clear text"
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid username/password combination"
definitions:
  User:
    type: "object"
    required:
      - "username"
      - "firstName"
      - "lastName"
      - "email"
      - "password"
    properties:
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"